rename package to mofun (suggestion from Kutay ;)
