+++ README.md
@@ -1,1 +1,1 @@
-# functionaliseMOF
+# mofun
